refactor!: Move large parts of deserialized_responses to matrix-sdk-base
