+++ crates/matrix-sdk-base/src/lib.rs
@@ -25,6 +25,7 @@ pub use crate::{
 };
 
 mod client;
+pub mod deserialized_responses;
 mod error;
 pub mod media;
 mod rooms;
@@ -32,6 +33,7 @@ mod session;
 #[cfg(feature = "sliding-sync")]
 mod sliding_sync;
 pub mod store;
+pub mod sync;
 mod utils;
 
 pub use client::BaseClient;
