+++ crates/matrix-sdk-common/src/deserialized_responses.rs
@@ -1,47 +1,11 @@
-use std::{borrow::Borrow, collections::BTreeMap};
+use std::collections::BTreeMap;
 
 use ruma::{
-    api::client::{
-        push::get_notifications::v3::Notification,
-        sync::sync_events::{
-            v3::{Ephemeral, InvitedRoom, Presence, RoomAccountData, State},
-            DeviceLists, UnreadNotificationsCount as RumaUnreadNotificationsCount,
-        },
-    },
-    events::{
-        room::member::{
-            MembershipState, RoomMemberEvent, RoomMemberEventContent, StrippedRoomMemberEvent,
-            SyncRoomMemberEvent,
-        },
-        AnyGlobalAccountDataEvent, AnyRoomAccountDataEvent, AnySyncTimelineEvent, AnyTimelineEvent,
-        AnyToDeviceEvent,
-    },
+    events::{AnySyncTimelineEvent, AnyTimelineEvent},
     serde::Raw,
-    DeviceKeyAlgorithm, EventId, MilliSecondsSinceUnixEpoch, OwnedDeviceId, OwnedEventId,
-    OwnedRoomId, OwnedUserId, UserId,
+    DeviceKeyAlgorithm, OwnedDeviceId, OwnedEventId, OwnedUserId,
 };
 use serde::{Deserialize, Serialize};
-
-/// A change in ambiguity of room members that an `m.room.member` event
-/// triggers.
-#[derive(Clone, Debug, Default, Deserialize, Serialize)]
-pub struct AmbiguityChange {
-    /// Is the member that is contained in the state key of the `m.room.member`
-    /// event itself ambiguous because of the event.
-    pub member_ambiguous: bool,
-    /// Has another user been disambiguated because of this event.
-    pub disambiguated_member: Option<OwnedUserId>,
-    /// Has another user become ambiguous because of this event.
-    pub ambiguated_member: Option<OwnedUserId>,
-}
-
-/// Collection of ambiguioty changes that room member events trigger.
-#[derive(Clone, Debug, Default, Deserialize, Serialize)]
-pub struct AmbiguityChanges {
-    /// A map from room id to a map of an event id to the `AmbiguityChange` that
-    /// the event with the given id caused.
-    pub changes: BTreeMap<OwnedRoomId, BTreeMap<OwnedEventId, AmbiguityChange>>,
-}
 
 /// The verification state of the device that sent an event to us.
 #[derive(Clone, Debug, Deserialize, Serialize)]
@@ -121,38 +85,6 @@ impl From<TimelineEvent> for SyncTimelineEvent {
     }
 }
 
-#[derive(Clone, Debug, Default, Deserialize, Serialize)]
-pub struct SyncResponse {
-    /// The batch token to supply in the `since` param of the next `/sync`
-    /// request.
-    pub next_batch: String,
-    /// Updates to rooms.
-    pub rooms: Rooms,
-    /// Updates to the presence status of other users.
-    pub presence: Presence,
-    /// The global private data created by this user.
-    pub account_data: Vec<Raw<AnyGlobalAccountDataEvent>>,
-    /// Messages sent directly between devices.
-    pub to_device_events: Vec<Raw<AnyToDeviceEvent>>,
-    /// Information on E2E device updates.
-    ///
-    /// Only present on an incremental sync.
-    pub device_lists: DeviceLists,
-    /// For each key algorithm, the number of unclaimed one-time keys
-    /// currently held on the server for a device.
-    pub device_one_time_keys_count: BTreeMap<DeviceKeyAlgorithm, u64>,
-    /// Collection of ambiguity changes that room member events trigger.
-    pub ambiguity_changes: AmbiguityChanges,
-    /// New notifications per room.
-    pub notifications: BTreeMap<OwnedRoomId, Vec<Notification>>,
-}
-
-impl SyncResponse {
-    pub fn new(next_batch: String) -> Self {
-        Self { next_batch, ..Default::default() }
-    }
-}
-
 #[derive(Clone, Debug)]
 pub struct TimelineEvent {
     /// The actual event.
@@ -160,173 +92,6 @@ pub struct TimelineEvent {
     /// The encryption info about the event. Will be `None` if the event was not
     /// encrypted.
     pub encryption_info: Option<EncryptionInfo>,
-}
-
-#[derive(Clone, Debug, Default, Deserialize, Serialize)]
-pub struct Rooms {
-    /// The rooms that the user has left or been banned from.
-    pub leave: BTreeMap<OwnedRoomId, LeftRoom>,
-    /// The rooms that the user has joined.
-    pub join: BTreeMap<OwnedRoomId, JoinedRoom>,
-    /// The rooms that the user has been invited to.
-    pub invite: BTreeMap<OwnedRoomId, InvitedRoom>,
-}
-
-/// Updates to joined rooms.
-#[derive(Clone, Debug, Deserialize, Serialize)]
-pub struct JoinedRoom {
-    /// Counts of unread notifications for this room.
-    pub unread_notifications: UnreadNotificationsCount,
-    /// The timeline of messages and state changes in the room.
-    pub timeline: Timeline,
-    /// Updates to the state, between the time indicated by the `since`
-    /// parameter, and the start of the `timeline` (or all state up to the
-    /// start of the `timeline`, if `since` is not given, or `full_state` is
-    /// true).
-    pub state: State,
-    /// The private data that this user has attached to this room.
-    pub account_data: Vec<Raw<AnyRoomAccountDataEvent>>,
-    /// The ephemeral events in the room that aren't recorded in the timeline or
-    /// state of the room. e.g. typing.
-    pub ephemeral: Ephemeral,
-}
-
-impl JoinedRoom {
-    pub fn new(
-        timeline: Timeline,
-        state: State,
-        account_data: Vec<Raw<AnyRoomAccountDataEvent>>,
-        ephemeral: Ephemeral,
-        unread_notifications: UnreadNotificationsCount,
-    ) -> Self {
-        Self { unread_notifications, timeline, state, account_data, ephemeral }
-    }
-}
-
-/// Counts of unread notifications for a room.
-#[derive(Copy, Clone, Debug, Default, Deserialize, Serialize)]
-pub struct UnreadNotificationsCount {
-    /// The number of unread notifications for this room with the highlight flag
-    /// set.
-    pub highlight_count: u64,
-    /// The total number of unread notifications for this room.
-    pub notification_count: u64,
-}
-
-impl From<RumaUnreadNotificationsCount> for UnreadNotificationsCount {
-    fn from(notifications: RumaUnreadNotificationsCount) -> Self {
-        Self {
-            highlight_count: notifications.highlight_count.map(|c| c.into()).unwrap_or(0),
-            notification_count: notifications.notification_count.map(|c| c.into()).unwrap_or(0),
-        }
-    }
-}
-
-#[derive(Clone, Debug, Deserialize, Serialize)]
-pub struct LeftRoom {
-    /// The timeline of messages and state changes in the room up to the point
-    /// when the user left.
-    pub timeline: Timeline,
-    /// Updates to the state, between the time indicated by the `since`
-    /// parameter, and the start of the `timeline` (or all state up to the
-    /// start of the `timeline`, if `since` is not given, or `full_state` is
-    /// true).
-    pub state: State,
-    /// The private data that this user has attached to this room.
-    pub account_data: RoomAccountData,
-}
-
-impl LeftRoom {
-    pub fn new(timeline: Timeline, state: State, account_data: RoomAccountData) -> Self {
-        Self { timeline, state, account_data }
-    }
-}
-
-/// Events in the room.
-#[derive(Clone, Debug, Default, Deserialize, Serialize)]
-pub struct Timeline {
-    /// True if the number of events returned was limited by the `limit` on the
-    /// filter.
-    pub limited: bool,
-
-    /// A token that can be supplied to to the `from` parameter of the
-    /// `/rooms/{roomId}/messages` endpoint.
-    pub prev_batch: Option<String>,
-
-    /// A list of events.
-    pub events: Vec<SyncTimelineEvent>,
-}
-
-impl Timeline {
-    pub fn new(limited: bool, prev_batch: Option<String>) -> Self {
-        Self { limited, prev_batch, ..Default::default() }
-    }
-}
-
-/// Wrapper around both MemberEvent-Types
-#[allow(clippy::large_enum_variant)]
-#[derive(Clone, Debug, Deserialize, Serialize)]
-pub enum MemberEvent {
-    Sync(SyncRoomMemberEvent),
-    Stripped(StrippedRoomMemberEvent),
-}
-
-impl MemberEvent {
-    /// The inner Content of the wrapped Event
-    pub fn original_content(&self) -> Option<&RoomMemberEventContent> {
-        match self {
-            MemberEvent::Sync(e) => e.as_original().map(|e| &e.content),
-            MemberEvent::Stripped(e) => Some(&e.content),
-        }
-    }
-    /// The sender of this event.
-    pub fn sender(&self) -> &UserId {
-        match self {
-            MemberEvent::Sync(e) => e.sender(),
-            MemberEvent::Stripped(e) => e.sender.borrow(),
-        }
-    }
-    /// The ID of this event.
-    pub fn event_id(&self) -> Option<&EventId> {
-        match self {
-            MemberEvent::Sync(e) => Some(e.event_id()),
-            MemberEvent::Stripped(_) => None,
-        }
-    }
-    /// The Server Timestamp of this event.
-    pub fn origin_server_ts(&self) -> Option<MilliSecondsSinceUnixEpoch> {
-        match self {
-            MemberEvent::Sync(e) => Some(e.origin_server_ts()),
-            MemberEvent::Stripped(_) => None,
-        }
-    }
-
-    /// The membership state of the user
-    pub fn membership(&self) -> &MembershipState {
-        match self {
-            MemberEvent::Sync(e) => e.membership(),
-            MemberEvent::Stripped(e) => &e.content.membership,
-        }
-    }
-
-    /// The user id associated to this member event
-    pub fn user_id(&self) -> &UserId {
-        match self {
-            MemberEvent::Sync(e) => e.state_key(),
-            MemberEvent::Stripped(e) => &e.state_key,
-        }
-    }
-}
-
-/// A deserialized response for the rooms members API call.
-///
-/// [GET /_matrix/client/r0/rooms/{roomId}/members](https://matrix.org/docs/spec/client_server/r0.6.0#get-matrix-client-r0-rooms-roomid-members)
-#[derive(Clone, Debug, Default, Deserialize, Serialize)]
-pub struct MembersResponse {
-    /// The list of members events.
-    pub chunk: Vec<RoomMemberEvent>,
-    /// Collection of ambiguity changes that room member events trigger.
-    pub ambiguity_changes: AmbiguityChanges,
 }
 
 #[cfg(test)]
